Get a list of available data sources

Get a list of available data sources

Usage

mi_sources(level, year = NULL, limit = 2500)

Arguments

level

a character string specifying the NUTS level ("0", "1", "2", or "3"). You can also always check valid NUTS levels using mi_nuts_levels.

year

an integer of length 1, specifying the year. Optional.

limit

An integer specifying the maximum number of results to return. Default is 2500. This default should be enough for most uses, as it is well above the number of NUTS 3 regions in the EU. The maximum limited by the API is 10000.

Value

a tibble of sources with the following columns:

source_name: name of the data source

short_description: short description of the data source

description: description of the data source

Examples

# \donttest{
# get up to 10 sources for NUTS level 3
mi_sources("3", limit = 10)
#> # A tibble: 10 × 3
#>    source_name    short_description      description                            
#>    <chr>          <chr>                  <chr>                                  
#>  1 DEMO_R_D3AREA  "Area of regions"      Area by NUTS 3 region (ESTAT)          
#>  2 PROJ_19RAASFR3 "Fertility assumption" Assumptions for fertility rates by age…
#>  3 PROJ_19RAASMR3 "Death assumptions"    Assumptions for probability of dying b…
#>  4 BD_HGNACE2_R3  "Business demography " Business demography and high growth en…
#>  5 BD_SIZE_R3     "Business demography " Business demography by size class and …
#>  6 CENS_11DWOB_R3 "Dwellings by groups"  Conventional dwellings by occupancy st…
#>  7 eter           "Higher Ed Institutes" Count of higher education institutions…
#>  8 CRIM_GEN_REG   "Crimes by region"     Crimes recorded by the police by NUTS …
#>  9 DEMO_R_MAGEC3  "Deaths by goups"      Deaths by age group, sex and NUTS 3 re…
#> 10 DEMO_R_DEATHS  "Deaths by region"     Deaths (total) by NUTS 3 region (ESTAT)

# get all sources for NUTS level 3 and year 2020
mi_sources("3", year = 2020)
#> # A tibble: 19 × 3
#>    source_name    short_description      description                            
#>    <chr>          <chr>                  <chr>                                  
#>  1 PROJ_19RAASFR3 "Fertility assumption" Assumptions for fertility rates by age…
#>  2 PROJ_19RAASMR3 "Death assumptions"    Assumptions for probability of dying b…
#>  3 BD_HGNACE2_R3  "Business demography " Business demography and high growth en…
#>  4 BD_SIZE_R3     "Business demography " Business demography by size class and …
#>  5 eter           "Higher Ed Institutes" Count of higher education institutions…
#>  6 DEMO_R_MAGEC3  "Deaths by goups"      Deaths by age group, sex and NUTS 3 re…
#>  7 DEMO_R_DEATHS  "Deaths by region"     Deaths (total) by NUTS 3 region (ESTAT)
#>  8 BD_ENACE2_R3   "Employer demography"  Employer business demography by NACE R…
#>  9 DEMO_R_FIND3   "Fertility indicators" Fertility indicators by NUTS 3 region …
#> 10 ookla          "Ookla internet speed" Fixed broadband and mobile (cellular) …
#> 11 DEMO_R_FAGEC3  "Live births by group" Live births by age group of the mother…
#> 12 DEMO_R_BIRTHS  "Live births, regions" Live births (total) by NUTS 3 region (…
#> 13 pm25           "Air particulates"     PM2.5 Satellite-derived particulate ma…
#> 14 DEMO_R_GIND3   "Population change - " Population change - Demographic balanc…
#> 15 DEMO_R_D3DENS  "Population density"   Population density by NUTS 3 region (E…
#> 16 PROJ_19RALEXP3 "Life expectancy"      Projected life expectancy by age (reac…
#> 17 REGION_SOCIAL  "Indicators Soc & Env" Regional Social and Environmental indi…
#> 18 ghs_smod       "Settlement type"      Satellite-based settlement types based…
#> 19 REGION_ST      "Short-term stats"     Short-term regional statistics (OECD)  

# }